ポリゴン削除部分を別クラスに分離
